{{cookiecutter.bot_name}}/requirements.txt
botbuilder-core>=4.7.0
aiohttp
httpx

coco-microsoft-bot-framework